move SigProcVariable to detail
